Revert "Acertando o tipo de retorno para uint8_t que é o tipo do ID de tabela."
This reverts commit 44461ba11acafc853158997a63db37f587217c8f.

+++ scratch/svelte/infrastructure/backhaul-controller.h
@@ -156,7 +156,7 @@ protected:
    * \param slice The slice ID.
    * \return The pipeline table for this slice.
    */
-  uint8_t GetSliceTable (SliceId slice) const;
+  uint16_t GetSliceTable (SliceId slice) const;
 
   /**
    * Notify this controller of a new bearer context created.
